Fix issue with Virtual vs. real URLs.
In r273854 a regression was introduced in which
HostZoomMapImpl::SetZoomLevelForWebContents() used a virtual url for
setting page zoom, when the real url is required. This CL fixes that
by calling GetLastCommittedEntry() on the WebContents navigation
controller, instead of calling it directly on the WebContents; the
latter returns a virtual url.

BUG=379622

Review URL: https://codereview.chromium.org/301243018