Reset avatar image when the avatar policy becomes unset
When the avatar policy is cleared, the user can pick an avatar image
again. Until the user does this, Chrome has two options:
1) keep using the image that had most recently been set through policy
2) revert to a default image

Chrome's current behavior is (1). This CL changes it to (2), which may be
easier to understand for users and is consistent with the behavior of the
wallpaper policy.

BUG=380422
TEST=Updated browser test

Review URL: https://codereview.chromium.org/315213002