Add ability to set the border insets of a box layout.
This CL makes it possible to give a border insets to box layout which
overrides the constructor settings for border spacing. This lets the
client have more control on the insets of its child views.

This is a precursor to https://codereview.chromium.org/307333002/.

BUG=349727

Review URL: https://codereview.chromium.org/310753002