[NaCl SDK] Don't set minimum_chrome_version for most examples.
Most examples work with Chrome stable (currently M35), so don't require
apps to use a newer version of Chrome.

No examples need a later version of Chrome, but a few use features that
just made it to stable. In this case, we allow the minimum version to be
specified in example.dsc.

BUG=none
R=sbc@chromium.org, noelallen@chromium.org

Review URL: https://codereview.chromium.org/318823002