Add list of Shared Modules to App Info Dialog
Add a list of Shared Modules to the App Info Dialog, along with an
'About' link next to each one that opens the shared module's options
page. This allows shared modules to display licenses, which meets a
legal requirement explained more in the bug.

BUG=364681

Review URL: https://codereview.chromium.org/314613002