Files.app: Omit first draw without metadata in file list
Before this patch, Files.app was draw the filelist twice.
1) After fetching list of file entry, draw the list without metadata.
2) After fetching the metadata, draw the list with metadata.

This way was because the metadata fetch had been slow, so we drew the list without metadata at the first time. But this took long time in total since drawing was twice (). And sometimes #2 was not happen and caused the issue since #2 was triggered by sort.

Now fetching metadata is enough fast, and we can combine them. This patch omit the first draw #1. This improves the performance, and fixes crbug.com/375575.

BUG=375575, 367123
TEST=filelist was drawn correctly
R=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/301943002