binary_size: Check debug data format before starting.
If the debug data format is DWARF4 and binutils is not new enough
the output will be more or less useless (no files will be detected).
Detect that case early instead of having the user wait for many hours
for something that can't be used.

BUG=370378

Review URL: https://codereview.chromium.org/302633003