Don't assume WebContents has a LastCommittedEntry when setting zoom.
In HostZoomMapImpl::SetZoomLevelForWebContents() we currently make the
assumption that the WebContents passed in will always have a non-null
last-committed navigation entry, but this leads to failures on the
fuzzing tests. This CL removes that assumption, and ignores attempts to
set zoom if no such entry exists.

BUG=382320

Review URL: https://codereview.chromium.org/320863002