Make domdistiller.js not modify `window`
Currently when we inject javascript into the page, we clobber a whole
bunch of names on the page's window object.

This moves the included javascript into the anonymous function so that
defining a function/var doesn't add that name to the window object and
also passes a new object with the window as prototype as the 'window'
variable to that function so that if it explicitly modifies window it
will modify the new object instead.

NOTRY=true

Review URL: https://codereview.chromium.org/320113002