Fix the ComboBox Menu in views_examples_with_content_exe when chromeos=1
Since before M32, attempting to show the dropdown menu in views_examples
in a chromeos=1 build would crash. The problem is that there is no Aura
DispatcherClient set in the example exe. Setting this is usually the
responsibility of the ash Shell.

This change adds a dummy wm::NestedAcceleratorDelegate for use in the
example exe on ChromeOS.

BUG=379624

Review URL: https://codereview.chromium.org/313543002