Allow "freezing" NativePages to spare up memory.
If the user opens many tabs and navigates to a NativePage in each tab,
the NativePages may use enough memory to cause a Java out-of-memory error.
This adds a method Tab.freezeNativePage() which destroys the current
native page and replaces it with a FrozenNativePage, which allows querying
getTitle(), getUrl(), etc, but can't be displayed.

BUG=359573

Review URL: https://codereview.chromium.org/322253002