Add limitation that LoadNexeAndStart must be called on the main thread.
Now we're switching LoadNexeAndStart from synchronous implementation to
asynchronous one. Under PPAPI implementation, the async callbacks are
handled on main thread.
To keep the similicity of the thread model, this CL introduces the limitation
that Plugin::LoadNexeAndStart must be called on the main thread.

TEST=Ran browser_tests --gtest_filter=NaCl* locally and trybots.
BUG=333950
R=teravest@chromium.org

Review URL: https://codereview.chromium.org/321053004