Ensure ephemeral apps have dock icons on mac
App shims are now created for ephemeral apps in the user data directory
to ensure that they have dock icons. App shims are not copied to the
Applications folder as ephemeral apps should be hidden from the OS.

BUG=375027
TEST=Launch an app ephemerally and ensure it has a dock icon, but does
not appear in Spotlight search results. Install the app fully and
ensure it now has both a dock icon and appears in Spotlight.

Review URL: https://codereview.chromium.org/311293002