Support minimum gesture bounds in GestureProvider
Some touch devices report either zero or unreasonably small touch sizes. While
this doesn't present a problem for gesture detection, it does affect the
behavior of the touch-derived gestures. For example, tap disambiguation is
prevented if the tap gesture's size is sufficiently small. As a workaround,
provide a platform-specific minimum gesture bounds length in GestureProvider.
On Android, this value is the same as was used before the unified gesture
detector, and on Aura we simply use half the default touch size.

BUG=379876

Review URL: https://codereview.chromium.org/321563002