Make content_extractor support a multiple-url request
This adds the --urls flag which accepts a space-separated list of urls
to distill.

This adds a pretty straightforward approach to having multiple of these
requests happening at the same time.

Once all requests are finished, they will be printed (in the order that
they appeared in --urls). If printing binary, each serialized protobuf
will be preceded by its size (as a varint32).

TBR=ben@

Review URL: https://codereview.chromium.org/276553002