Password manager internals page: Add logging for parsed forms
The PasswordAutofillAgent::SendPasswordForms method is called
both when HTML forms are just parsed, and also when they are rendered.
The former is the moment when the password manager first hears about
forms, and notes their presence not only for autofill, but also
for matching against credentials submitted later.

However, I missed the second use-case (matching submitted credentials)
when I put logging there for the internals page, and so currently
the progress is only logged when the forms are rendered.

This CL should change this: logs are now also sent when the forms have just been parsed.

I already saw issues (e.g., http://crbug.com/367768) where such additional logging would have sped up the investigation.

There are no tests, because we don't which logs are generated when.

BUG=347927

Review URL: https://codereview.chromium.org/323423006