Don't fetch histograms from the same process.
This change makes HistogramController not fetch any
histograms from a "child" process when that process
is the same as the current process. In reality, this
could be the case on Android where the browser
process is actually also the GPU process.

See associated bug for the specific issue this is
actually fixing.

Also updates the profiler child process code to
use the same logic (it was not affected by the
bug, though).

BUG=370208

Review URL: https://codereview.chromium.org/328823002