cc: LTHI decides which tree is used for commit.
LTH should not do anything special based on whether impl-side painting
is enabled or not. It should just get a tree from LTHI and commit to it.
This will allow us to get rid of pending tree when rasterization happens
synchronously on the compositor thread.

BUG=383157

Review URL: https://codereview.chromium.org/321223005