veatest: Create and destroy all encoders on one thread.
In Chrome, encoders are created and destroyed on the GPU ChildThread.
In the test though, we create all of them on separate threads. This poorly
simulates the real-world scenario, and may also cause issues. Switching to
one, main "ChildThread", on which to create the encoders, will help us
ensure they are capable of sharing it properly. Moreover, having multiple
"ChildThread"s would not be a good idea on X11, because calling into it
from multiple threads is not allowed and does not work.

BUG=None
TEST=Run the test.

Review URL: https://codereview.chromium.org/333883002