Revert of Blacklists virtualCamera.ax (https://codereview.chromium.org/325323003/)
Reason for revert:
No crashes have been found after 37.0.2046.2 and Canary on Windows is now on 37.0.2051.2, hence it looks good. Not sure if was the blacklisting or the "calling CoInitializeEx twice" patch which did the trick here.

I would like to revert the blacklisting patch to see if we can cope with the CoInitializeEx patch only. It would be great if we could avoid adding another DLL to the blacklist.

Original issue's description:
> Blacklists virtualCamera.ax.
> 
> Tries to avoid crash in media::CoreAudioUtil::CreateDeviceEnumerator by blacklisting a DLL which accounts for a large amount of crashes we see in CreateDeviceEnumerator.
> 
> BUG=383715, 378465
> 
> Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=276946

TBR=tommi@chromium.org,robertshield@chromium.org
NOTREECHECKS=true
NOTRY=true
BUG=383715, 378465

Review URL: https://codereview.chromium.org/335073002