Change HistoryService::QueryURL to use CancelableTaskTracker
The callback no longer receive an HistoryService::Handle as first parameter
and the caller should use the returned base::CancelableTaskTrasker::TaskId
to cancel a request early.

Instead of passing a CancelableRequestConsumer to the query, the caller
pass a base::CancelableTaskTracker. The lifetime of the callback can be
associated to the request by using base::Passed().

Remove unused types and fix usage of the values passed to the result
callback (check success before using the URLRow or the VisitVector).

BUG=371818
TBR=asargent@chromium.org
TBR=bauerb@chromium.org
TBR=blundell@chromium.org
TBR=gavinp@chromium.org
TBR=pkasting@chromium.org
TBR=shess@chromium.org
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/333843003