Refactor tool-bar event handlers.
The tool-bar is being phased out as part of the apps v2 work, but as
part of the transition, the functionality it provides needs to be shared
between it and the window frame. Step one is to disentangle some of the
event handling code from client_session.js and toolbar.js.

NOTRY=true

Review URL: https://codereview.chromium.org/336293003