[Android WebView] Fix hang in popup window flow.
It's possible that we end up hanging the browser
process waiting for the response to a popup window
due to a bug in the AwContents hand-over when swapping
out the popup content into the user-supplied WebView.

This patch ensures that we don't end up with two AwContents
instances pointing to the same java AwContentsClientBridge
java object during deletion.

BUG=

Review URL: https://codereview.chromium.org/305823005