Don't call AutocompleteInput::Parse() on a background thread, part 1.
Instead of creating the exact match in HistoryURLProvider::DoAutocomplete(),
pass it in in the HistoryURLProviderParams struct.

This also uses this match as the "can't find DB, fallback" match in Start(); I
don't see any disadvantage to creating that match on the results of running
fixup, instead of creating a pre-fixup match like before.

This also splits out CanPromoteMatchForInlineAutocomplete() from
PromoteMatchForInlineAutocomplete().  I'll be converting
PromoteMatchForInlineAutocomplete() into a different function in the next part
of this change, and this happens to make the calling convention a little more
sane anyway.

Also removes a few comments from the .cc file that are in the .h file already.

BUG=376199
TEST=none
R=mpearson@chromium.org

Review URL: https://codereview.chromium.org/336173005