app-shell: Introduce a ShellRendererMainDelegate.
It may be necessary to create custom filters/observers etc. on the renderer
side in athena (e.g., for virtual keyboard support). So allow athena to install
a delegate to be created in the renderer process. The delegate is notified
when the RenderThread is initialized, or when new RenderViews are created.

BUG=380215
R=jamescook@chromium.org, oshima@chromium.org

Review URL: https://codereview.chromium.org/331323008