[fsp] Cleanup handling errors for operation requests.
This CL removes redundant internal functions which are identical, and replaces
them with one.

Moreover, the new general purpose error API function, passes now the entire
value to the request manager. This is because, very soon, we are going to
pass extra information with every error and success responses and we want to
have access to them in the RequestManager.

NOTRY=true
TEST=unit_tests, browser_tests: *FileSystemProvider*
BUG=373165, 384201

Review URL: https://codereview.chromium.org/335753004