Force resetting the VK insets when overscroll mode is disabled.
The login and lock screens can toggle overscroll mode for the virtual keyboard without triggering a rebuild of the VK.  Thus, it is possible to be in an inconsistent state between keyboard bounds change notifications.  This patch forces any insets to be reset if overscroll mode is not enabled.

BUG=384981

Review URL: https://codereview.chromium.org/341623002