MSE: Reduce spurious discontinuities caused by partial append window start overlaps
If a frame overlaps appendWindowStart, partial append window filtering
is supported for the track, and the frame is appended to the track
buffer, this change updates the track buffer's last frame duration with
the frame's original duration instead of the possibly trimmed duration.
If the trimmed duration were used, then the discontinuity detection
logic could detect a discontinuity on the next frame for the track where
there would otherwise not be a discontinuity.

BUG=381114
R=acolwell@chromium.org
TEST=FrameProcessorTest.PartialAppendWindowFilterNoDiscontinuity

Review URL: https://codereview.chromium.org/344513002