Changes routing of input events in view manager
Events are now sent to the window manager (assumed to be at connection
id 1 for now). The window manager can then do what it wants with
them. For now I made it bounce the event back to the server so
everything works as it did.

Until we get FIFO across pipes I'm putting these messages on the
ViewManager/ViewManagerClient interface. That'll change.

BUG=384433
TEST=covered by tests
R=ben@chromium.org

Review URL: https://codereview.chromium.org/338353006