Remove TraceStringWrapper::operator const char*()
Remove the operator to avoid misuse like the following:
  TRACE_EVENT1("a", "b", "c",
     condition ? TRACE_STR_COPY(short_live_string) : "literal");
In the above case, the argument value is actually const char*, not
TraceValueWithCopy because it is automatically converted back to
const char* with the const char* operator.

Found no errornous usages in chromium, but found one error in WebKit.

R=dsinclair
BUG=382789 (perhaps the reason)

Review URL: https://codereview.chromium.org/338653005