Battery Status API: end-to-end browsertests and default values.
This patch adds end-to-end testing for the Battery Status API.
Also a small fix in battery_status_service to return default
values if battery_status_manager fails to start (e.g. no battery
or no implementation). That way the promise on the blink side
will resolve with default values as according to the specification.

BUG=122593, 360068

Review URL: https://codereview.chromium.org/334223003