Add a dependency on libelf-dev:i386.
This is needed for https://codereview.chromium.org/341573003/
On Lucid, libelf-dev:i386 couldn't be installed in parallel to the 64bit
version. So nacl has a hack to remove -m32 flags from a target that depends
on libelf, to make a target_arch=ia32 on a 64bit system with multilibs work.

In precise, libelf-dev:i386 can be installed in parallel, so do so as a
prerequisite for removing this hack.

BUG=368384
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/347663002