Override HOME dir for gsutil when running on cros device
Gsutil tries to create a dir (for keeping track of resumable downloads)
under home dir. On ChromeOS this is, given that the tests must be run
by root, read-only '/root/'. This causes gsutil commands to fail. To
avoid this, override HOME dir to writable '/home/chromeos-test'.

(this will also require gsuitl to be configured with the same HOME dir)

BUG=359293,386416

Review URL: https://codereview.chromium.org/334603004