Enable pointer warping with multiple displays under Ozone
Mouse events generated from Ozone do not have a NativeEvent associated
with them and so they need to go through the non native event based flow
in the multi-monitor warping code. With this change it is now possible
to move the pointer from one display to the next under Ozone.

TEST=On link_freon attach a second display and move the pointer from
one display to the other.
BUG=361144

Review URL: https://codereview.chromium.org/340773003