[Cast] Fix log_deserializer BasePacketEvent merge bug.
The bug is that if we are merging BasePacketEvent A into B, and A had
a valid packet size but B did not (i.e. because it only has receiver
events and therefore no packet size) then the merged result would take
on the packet size of B which is invalid.

The fix is to take the max of the sizes of A and B and set it on the
merged result.

Review URL: https://codereview.chromium.org/331383006