Calculate the tabstrip min width based on the actual tabs in the strip.
The maximum minimum width is the width required to show 20 tabs at minimum size.
In my testing, that resulted in an overall window width of 440 pixels.  This is
narrow enough that I doubt this minimum will kick in very often, and when it
does, it won't be dramatically larger than the minimum width the toolbar already
enforces.

BUG=62510
TEST=Open 20 tabs.  Shrink Chrome window to minimum width and ensure all tabs are still visible.

Review URL: https://codereview.chromium.org/343603003