Don't call AutocompleteInput::Parse() on a background thread, part 2.
This changes HistoryURLProviderParams from holding an ACMatches object to
holding a HistoryMatches object; the HistoryMatches will no longer be fixed up
in DoAutocomplete(), but rather after it returns control to the UI thread.

The majority of this change is mechanical, but some nontrivial changes have to
be made to DoAutocomplete() to account for how the "promote" action it could
previously take must also happen on the UI thread.  Therefore the |params|
object also has to gain a few members to allow the post-DoAutocomplete() code to
do promotion correctly.

BUG=376199
TEST=none
R=mpearson@chromium.org

Review URL: https://codereview.chromium.org/333273009