Use a TestBrowserThreadBundle in the DeviceSettingsTestBase.
Lots of policy tests inherit from that class to reuse common setup code.
The recent work on cloud policy for extensions wired a new class to some
of the existing ones that posts a cleanup task to the IO thread, and
DeviceSettingsTestBase wasn't mocking that thread; so these tests started
leaking at shutdown.

The TestBrowserThreadBundle sets up the UI and FILE threads that the tests
were expecting as well as the other common threads, which makes these tests
more robust.

TBR=pastarmovj@chromium.org
BUG=None

Review URL: https://codereview.chromium.org/348713004