athena: Add support for virtual keyboard.
The virtual keyboard used on Chrome OS (in ash) is provided by an
extension. For athena, we cannot yet use the extension version of the
keyboard. As a short-term workaround, this patch introduces a web-ui
version of the keyboard for initial use. This patch provides a
simplistic implementation of some of the virtualKeyboardPrivate API
(sendKeyEvent, hideKeyboard) over WebUI.

This patch installs a webui controller (VKWebUIController) for the
keyboard URL (chrome://keyboard), and a message-handler (VKMessageHandler)
to process the incoming messages from the keyboard. These are installed
in the browser process. In the renderer process, this patch installs a
RenderViewObserver implementation (VirtualKeyboardBindingsImpl) which
provides the necessary bindings for the virtualKeyboardPrivate API
(VKBindings), and generates the appropriate webui messages by calling
chrome.send(). This lets us use the same HTML/JS/Polymer version of the
keyboard as in Chrome OS under ash/chrome.

For the keyboard to be completely functional, an implementation of
onTextInputBoxFocused will also be necessary. That will be provided in a
follow up CL.

BUG=380125
R=ben@chromium.org, bshe@chromium.org, oshima@chromium.org

Review URL: https://codereview.chromium.org/335793005