[Mac] Misc fixes for the new avatar bubble
In this CL:
- crbug.com/343688: when editing the profile name but not changing it, the text field would
freeze. this is because it would remain enabled, but hidden, and steal the button click
(because it used to be a first responder)
- add a separator between the "not you" and "lock" bottom option buttons, to
match the mocks. I added this as a vertical separator to the BaseBubbleController itself, 
so that other classes can use it in the future.
- center the profile name correctly. Because of the pencil icon, the name would be shifted
slightly to the left
- the "change photo" overlay should be more opaque
- the bubble bottom corners were overlapped by the option buttons, and didn't appear rounded.

BUG=343688

Review URL: https://codereview.chromium.org/323143004