Add support for fixed-sized arrays in mojom
This adds support for the syntax:

  typename[20] name;

in mojom files to specify an array of 'typename' containing exactly 20
entries. The serialization is the same as 'typename[] name' and the length
of the array is checked at message validation time.

This implementation does not support nested fixed arrays as in:

  typename[3][4] name;

but does handle fixed size arrays inside structs, etc.

Review URL: https://codereview.chromium.org/340663002