Keep track of paths for fallback fonts in sandbox_ipc_linux
Similar to HandleFontMatchRequest we need to keep track of the
path of a font that was found when searching for a character fallback.
Otherwise the renderer cannot instantiate them later through
Skia's SkFontConfigInterface. Depends on Blink side change that extends
WebFallbackFont by fontconfigInterfaceId (r176494).

BUG=285303
R=piman,eae

Review URL: https://codereview.chromium.org/336153003